+++ docs/change-log.md
@@ -4,5 +4,9 @@
 
 # 0.2.0 (2025-01-24)
 ## Dev
-- setup mkdocs
-- take llm models and embedding models dependencies out of the main package dependencies.
+- Setup mkdocs
+- Take llm models and embedding models dependencies out of the main package dependencies.
+
+# 0.3.0 (2025-02-05)
+## models
+- Add support for huggingface llm models `LLMModel`